Fixes saltstack installation issue
This fixes the following error which we get currently:

amazon-ebs:  * ERROR: https://repo.saltstack.com/apt/ubuntu/16.04/amd64/2018.3/SALTSTACK-GPG-KEY.pub failed to download to /tmp/salt-gpg-2QoGbsZu.pub
amazon-ebs:  * ERROR: Failed to run install_ubuntu_stable_deps()!!!

On debugging this, it seems that public key doesn't exist for the
version 2018.3. But it seems to work for the version 2019.2.